Move all singleton ops to scala.compiletime.ops package
As suggested in code review
