--- library/src/scala/compiletime/boolean/package.scala
@@ -1,8 +1,0 @@
-package scala.compiletime
-
-package object boolean {
-  type ![X <: Boolean] <: Boolean
-  type ^[X <: Boolean, Y <: Boolean] <: Boolean
-  type &&[X <: Boolean, Y <: Boolean] <: Boolean
-  type ||[X <: Boolean, Y <: Boolean] <: Boolean
-}
