+++ library/src/scala/compiletime/package.scala
@@ -63,7 +63,4 @@ package object compiletime {
    *      }
    */
   type S[N <: Int] <: Int
-
-  type ==[X <: AnyVal, Y <: AnyVal] <: Boolean
-  type !=[X <: AnyVal, Y <: AnyVal] <: Boolean
 }
